requirements.txt
psycopg2-binary
pyobjc
pyinstaller